Update codeql-python-jupyter-analysis.yml
Updated actions/checkout to v3 and actions/setup-python to v4 to use the latest versions of these GitHub Actions, which include bug fixes, security patches, and new features.

Combined installation of dependencies (like nbconvert) into a single step to improve efficiency and readability.

Added github/codeql-action/autobuild@v2: This step automatically attempts to build the code, which is especially useful for dynamically typed languages or projects requiring a build process.
continue-on-error: true is used to allow the workflow to continue even if the autobuild step fails. This is helpful if the build step is non-blocking for the CodeQL analysis.